transition to georgia data
Prompt:

This repository contains workshop materials for two tidymodels workshops, intro
and advanced.

I'm working on the introductory workshop. I just made changes to the underlying
data from the forested package:

* The first version of the package just had data from Washington, called
 `forested`. The second release of the package also includes data from
Georgia.
 The `forested` data is the same as it was, and now is aliased as
`forested_wa`.
 The data from Georgia is called `forested_ga`. The Georgia data has the same
 structure, except is missing the `northness` variable due to issues with the
 source raster. `forested_ga` also includes a new variable `county`
 (#18, h.t. `@graysonwhite`).

* `forested` and its alias `forested_wa` gained a new variable, `county` (#14).

forested is my own package. I have this developmental version installed; it
will be on CRAN by the time participants see it.

I would like to transition the introductory workshop to use the `forested_ga`
data instead of the `forested` data. Could you make that happen? Update the
slides starting with `intro-` in `slides/` as well as `classwork/` (if needed). Note that the package, the dataset, and the target variable were called `forested`, so you can't just find and replace on that. Both the package and target variable are still called `forested`.